requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
requests==2.32.3
tflite-runtime==2.14.0
urllib3==2.2.2
Werkzeug==3.0.3
